fix: norm, split unused map to debug_minimap.c
