+++ Makefile
@@ -6,7 +6,7 @@ SRC_DIR = src
 OBJ_DIR = objs
 
 SRCS = $(addprefix $(SRC_DIR)/,\
-main.c draw.c hooks.c init.c oper.c parse.c render.c utils.c raycast.c free.c debug.c line.c pixel.c minimap.c mouse.c)
+main.c draw.c hooks.c init.c oper.c parse.c render.c utils.c raycast.c free.c debug.c line.c pixel.c minimap.c mouse.c debug_minimap.c)
 
 OBJS = $(patsubst $(SRC_DIR)/%.c, $(OBJ_DIR)/%.o, $(SRCS))
